// packages/metadata/src/types/abstract.ts
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSON Schema file,
 * and run generate-types to regenerate this file.
 */

/**
 * Schema for _abstract.yml — abstract text.
 */
export interface AbstractMetadata {
  /**
   * The abstract text.
   */
  abstract: string;
  [k: string]: unknown;
}